Refactor kstatus health check
